pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aio"
version = "0.1.0"
description = "AI-native presentation generator"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12.0",
    "jinja2>=3.1.2",
    "mistune>=3.0.2",
    "pyyaml>=6.0.1",
    "pygments>=2.15.0",
    "watchdog>=3.0.0",
    "httpx>=0.25.0",
    "click>=8.1.7",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.1",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "pyinstaller>=6.0.0",
    "types-pyyaml",
]
enrich = [
    "pillow>=10.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "cssutils>=2.7.0",
]
commands = [
    "pyperclip>=1.8.2",
]

[project.scripts]
aio = "aio.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aio = [
    "layouts/*.j2",
    "themes/**/*",
    "agent_commands/**/*.md",
    "themes/registry.json",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--tb=short"
